requirements.txt
streamlit
tomli
google-auth
google-cloud-aiplatform
bs4